Add offline support to ios
