+++ packages/image/__tests__/android/offline-image.android.test.js
@@ -1,0 +1,3 @@
+import shared from "../offline-image.native";
+
+shared();
