+++ packages/image/__tests__/ios/offline-image.ios.test.js
@@ -1,0 +1,3 @@
+import shared from "../offline-image.native";
+
+shared();
