+++ packages/image/src/utils/index.js
@@ -10,9 +10,9 @@ const appendParamsToQuery = (uriString, paramMap) => {
   paramMap.forEach(item => {
     if (
       !(
-        (item && finalUri.includes(`?${item.name}`)) ||
-        finalUri.includes(`&${item.name}`)
-      )
+        (item && (finalUri.includes(`?${item.name}`)) ||
+        finalUri.includes(`&${item.name}`))
+      ) && item.value
     ) {
       const separator = finalUri.includes("?") ? "&" : "?";
       finalUri = `${finalUri}${separator}${item.name}=${item.value}`;
